+++ dem/core/commands/add_cat_cmd.py
