+++ dem/core/commands/add_host_cmd.py
