+++ dem/core/commands/add_reg_cmd.py
