+++ dem/core/commands/assign_cmd.py
