+++ dem/core/commands/del_cat_cmd.py
