+++ dem/core/commands/del_host_cmd.py
