+++ dem/core/commands/del_reg_cmd.py
