+++ dem/core/commands/del_task_cmd.py
