+++ dem/core/commands/delete_cmd.py
