+++ dem/core/commands/info_cmd.py
