+++ dem/core/commands/init_cmd.py
