+++ dem/core/commands/list_cat_cmd.py
