+++ dem/core/commands/list_cmd.py
