+++ dem/core/commands/list_reg_cmd.py
