+++ dem/core/commands/rename_cmd.py
