+++ dem/core/commands/set_default_cmd.py
