+++ dem/core/commands/uninstall_cmd.py
